+++ relayer/chain/parachain/channels.go
@@ -2,15 +2,15 @@ package parachain
 
 import "github.com/snowfork/go-substrate-rpc-client/v3/types"
 
-type BasicOutboundChannelMessage struct {
+type OutboundChannelMessage struct {
 	Target  [20]byte
 	Nonce   uint64
 	Payload []byte
 }
 
+type BasicOutboundChannelMessage = OutboundChannelMessage
+
 type IncentivizedOutboundChannelMessage struct {
-	Target  [20]byte
-	Nonce   uint64
-	Fee     types.U256
-	Payload []byte
+	BasicOutboundChannelMessage
+	Fee types.U256
 }
